Add files via upload
There is a huge demand for used cars in the Indian Market today. As sales of new cars have slowed down in the recent past, the pre-owned car market has continued to grow over the past years and is larger than the new car market now. Cars4U is a budding tech start-up that aims to find footholes in this market.

The goal of the project is to explore, visualize the dataset, build a linear regression model to predict the prices of used cars, and generate a set of insights and recommendations that will help the business.

- The first part of this project is to explore, pre-process, and clean the data.
- The second deliverable would be to find the most important factors that affect the prices of cars using visualization.
- The last task is to build a predictive model to predict the prices of cars at Cars4U (The final package would include recommendations that would help the business).